3 files
cw_practice is an older version.  cw_words builds on that code. google-10000-english-usa.txt was a list of google searches at some point in the past.  It's not a list of English words per se, so it contains many vulgar words and weird abbreviations.  Please don't blame me.  It's just whatever people were searching for! It's not PG-13.